# Conflicts:
#	app/Exceptions/Handler.php
#	app/Http/Controllers/VoteController.php
#	app/Http/routes.php
#	app/Models/Vote.php

+++ app/Http/Controllers/VoteController.php
@@ -1,22 +1,48 @@
 <?php
-
 namespace App\Http\Controllers;
-
 use App\Models\Vote;
 use App\Models\User;
 use App\Http\Requests\VotesRequest;
-
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
-use Illuminate\Auth\Access\AuthorizationException;
-
 use DCN\RBAC\Traits\HasRoleAndPermission;
+use DCN\RBAC\Exceptions\PermissionDeniedException;
 use DCN\RBAC\Contracts\HasRoleAndPermission as HasRoleAndPermissionContract;
+
 
 class VoteController extends ApiController implements HasRoleAndPermissionContract
 {
     use HasRoleAndPermission;
+    
+    #TODO: Delete this after the authorization implement
+    public function __construct()
+    {
+        $users = User::all();
+        Auth::login($users[1]);
+    }
+    /**
+     * @param $votes array
+     * @return $data array
+     */
+    private function getMetaData($votes)
+    {
+        $data =[];
+        $i = 0;
 
+        foreach ($votes as $vote) {
+
+            if ($vote->is_saved) {
+                $data[$i]['data'] = $vote;
+                $data[$i]['_meta']['user'] = $vote->user()->first();
+                $data[$i]['_meta']['likes'] = $vote->likes()->count();
+                $data[$i]['_meta']['tags'] = $vote->tags()->count();
+                $data[$i]['_meta']['comments'] = $vote->comments()->count();
+                $i++;
+            }
+        }
+        return $data;
+    }
     /**
      * Display a listing of the resource.
      *
@@ -24,10 +50,14 @@ class VoteController extends ApiController implements HasRoleAndPermissionContra
      */
     public function index()
     {
-        $votes = Vote::all();
-        return $this->setStatusCode(200)->respond($votes);
-    }
+        $vote = new Vote();
+        if (!(Auth::user()->allowed('view.votes', $vote)))
+            throw new PermissionDeniedException('index');
 
+        $votes = Vote::all();
+        $data = $this->getMetaData($votes);
+        return $this->setStatusCode(200)->respond($data);
+    }
     /**
      * Store a newly created resource in storage.
      *
@@ -36,10 +66,13 @@ class VoteController extends ApiController implements HasRoleAndPermissionContra
      */
     public function store(VotesRequest $request)
     {
+        $vote = new Vote();
+        if (!(Auth::user()->allowed('create.votes', $vote)))
+            throw new PermissionDeniedException('create');
+
         $vote = Vote::create($request->all());
         return $this->setStatusCode(201)->respond($vote);
     }
-
     /**
      * Display the specified resource.
      *
@@ -49,9 +82,21 @@ class VoteController extends ApiController implements HasRoleAndPermissionContra
     public function show($id)
     {
         $vote = Vote::findOrFail($id);
-        return $this->setStatusCode(200)->respond($vote);
+
+        if (!(Auth::user()->allowed('view.votes', $vote)))
+            throw new PermissionDeniedException('view');
+
+        $user = $vote->user()->first();
+        $likeCount = $vote->likes()->count();
+        $tagCount = $vote->tags()->count();
+        $commentCount = $vote->comments()->count();
+
+        return $this->setStatusCode(200)->respond($vote, ['user' => $user,
+            'likes' => $likeCount,
+            'tags' => $tagCount,
+            'comments' => $commentCount]);
     }
-    
+
     /**
      * Update the specified resource in storage.
      *
@@ -63,10 +108,12 @@ class VoteController extends ApiController implements HasRoleAndPermissionContra
     {
         $vote = Vote::findOrFail($id);
 
+        if (!(Auth::user()->allowed('update.votes', $vote)))
+            throw new PermissionDeniedException('update');
+
         $vote->update($request->all());
         return $this->setStatusCode(200)->respond($vote);
     }
-
     /**
      * Remove the specified resource from storage.
      *
@@ -77,28 +124,44 @@ class VoteController extends ApiController implements HasRoleAndPermissionContra
     {
         $vote = Vote::findOrFail($id);
 
+        if (!(Auth::user()->allowed('delete.votes', $vote)))
+            throw new PermissionDeniedException('delete');
+
         $vote->delete();
-        return $this->setStatusCode(204)->respond();
+
+        if ($vote->trashed()) {
+            return $this->setStatusCode(204)->respond();
+        } else {
+            throw new \PDOException();
+        }
     }
 
     public function getUserVotes($userId)
     {
         $user = User::findOrFail($userId);
         $votes = $user->votes()->get();
+
+        $vote = $votes[0];
+        if (!(Auth::user()->allowed('view.votes', $vote)))
+            throw new PermissionDeniedException('index');
+
         if(!$votes){
             return $this->setStatusCode(200)->respond();
         }
         return $this->setStatusCode(200)->respond($votes, ['user' => $user]);
     }
-    
+
     public function getUserVote($userId, $voteId)
     {
         $user = User::findOrFail($userId);
-        $vote = $user->votes()->where('id',$voteId)->first();
+        $vote = $user->getVote($voteId);
+
+        if (!(Auth::user()->allowed('view.votes', $vote)))
+            throw new PermissionDeniedException('view');
+
         if(!$vote){
             throw (new ModelNotFoundException)->setModel(Vote::class);
         }
         return $this->setStatusCode(200)->respond($vote, ['user' => $user]);
-
     }
 }
